chore: fix hot loader in CRA
